+++ CHANGES.md
@@ -1,6 +1,10 @@
 Changes
 =======
 
+Version 2.0.1
+-------------
+- fixed an exploit
+
 Version 2.0.0
 -------------
 - implemented a new price calculation